src/main/java/com/bol/game/controllers/PlayerController.java
package com.bol.game.controllers;


import com.bol.game.pojos.Player;
import com.bol.game.repositories.PlayerRepository;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.http.MediaType;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RestController;

import javax.validation.Valid;

@RestController
public class PlayerController {

    private static final String PLAYERS_PATH = "/players";

    private final PlayerRepository playerRepository;

    @Autowired
    public PlayerController(PlayerRepository playerRepository) {
        this.playerRepository = playerRepository;
    }

    @PostMapping(path = PLAYERS_PATH,
            consumes = MediaType.APPLICATION_JSON_VALUE,
            produces = MediaType.APPLICATION_JSON_VALUE)
    public Player createPlayer(@Valid @RequestBody Player player) {
        return playerRepository.save(player);
    }

}
